# VMware: vmware_vm_inventory performance improvement