##### SUMMARY

Fixes: #56786

Signed-off-by: Abhijeet Kasurde <akasurde@redhat.com>


##### ISSUE TYPE 
- Bugfix Pull Request





##### COMPONENT NAME
lib/ansible/plugins/inventory/vmware_vm_inventory.py