@Akasurde 2019-12-18
@gparent Could you please try this PR and let me know? Thanks in advance.

For me - 

Before change
```
# time ansible-inventory -i vmware_vm_inventory.yml --list
...
2.29s user 0.21s system 1% cpu 2:50.43 total
```

After change 

```
# time ansible-inventory -i vmware_vm_inventory.yml --list
...
1.32s user 0.20s system 4% cpu 30.717 total
```