@goneri 2020-02-03
Hi @Akasurde,

I'm surprised because I actually see a performance drop:

- without the patch: 2678ms, 2788ms, 2465ms
- with the patch: 3404ms, 3247ms, 3261ms

Any idea what's going on?