@gparent 2020-02-03
I thought I had commented on this, I apologize. In my case the performance improved. I went from being unable to iterate my infrastructure to being able to assess about a hundred VM in a reasonable time once a folder was specified. I had taken numbers but it seems I never got to pasting them here.