@Akasurde 2020-03-09
Superseded by https://github.com/ansible-collections/vmware/pull/45/